Win cert changes (#46)
* Refactored code adding Windows cert store logic, including renaming IISU to WinIIS.

* Added PowerShell class to perform get-childitem from cert store

* Refactored code allowing multiple types of Cert Stores, including Win Cert and IIS (WebHosting) Cert Stores.

* Update generated README

* Fixed a problem adding certs to a cert store that had a space in the name (ie. Remote Desktop)
